[deep refactor] noop DRY componentize: card connector -> ConnectorButton
